task/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
cloudinary==1.44.1
Django==5.2.3
django-cloudinary-storage==0.3.0
djangorestframework==3.16.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==25.0
pillow==11.2.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
responses==0.25.7
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
